Moved the reading of rules and varrules (logic) into the unit base class. This means rules and varrules can be read from other file types using the same methods.
