+++ floodmodeller_api/units/_base.py
@@ -17,6 +17,11 @@ address: Jacobs UK Limited, Flood Modeller, Cottons Centre, Cottons Lane, London
 """ Holds the base unit class for all FM Units """
 
 from ..diff import check_item_with_dataframe_equal
+import pandas as pd
+from .helpers import (
+    split_10_char,
+    _to_float,
+)
 
 
 class Unit:
@@ -99,3 +104,128 @@ class Unit:
             diff=diff,
         )
         return (result, diff) if return_diff else result
+
+    # rules & varrules
+
+    def _read_logic(self, block):
+        logical_params = split_10_char(block[6])
+        self.max_movement_rate = logical_params[1]
+        self.max_setting = logical_params[2]
+        self.min_setting = logical_params[3]
+        self.gates = self._get_gates(self.ngates, block, gate_row=7)
+        rule_params = split_10_char(block[self._last_gate_row + 1])
+        self.nrules = int(rule_params[0])
+        self.rule_sample_time = _to_float(rule_params[1])
+        self.rules = self._get_logical_rules(
+            self.nrules, block, self._last_gate_row + 2
+        )
+        # Get time rule data set
+        nrows = int(split_10_char(block[self._last_rule_row + 1])[0])
+        data_list = []
+        for row in block[self._last_rule_row + 2 : self._last_rule_row + 2 + nrows]:
+            row_split = split_10_char(f"{row:<20}")
+            x = _to_float(row_split[0])  # time
+            y = row[10:].strip()  # operating rules
+            data_list.append([x, y])
+        self._last_time_row = self._last_rule_row + nrows + 1
+        rule_data = pd.DataFrame(data_list, columns=["Time", "Operating Rules"])
+        rule_data = rule_data.set_index("Time")
+        rule_data = rule_data["Operating Rules"]
+        self.time_rule_data = rule_data
+        # VARRULES (not always necessary)
+        if self._last_time_row + 1 < len(block):
+            if block[self._last_time_row + 1].strip() == "VARRULES":
+                varrule_params = split_10_char(block[self._last_time_row + 2])
+                self.nvarrules = int(varrule_params[0])
+                self.varrule_sample_time = _to_float(rule_params[1])
+                self.varrules = self._get_logical_rules(
+                    self.nvarrules, block, self._last_time_row + 3
+                )
+                # Get time rule data set
+                var_nrows = int(split_10_char(block[self._last_rule_row + 1])[0])
+                data_list = []
+                for row in block[
+                    self._last_rule_row + 2 : self._last_rule_row + 2 + var_nrows
+                ]:
+                    row_split = split_10_char(f"{row:<20}")
+                    x = _to_float(row_split[0])  # time
+                    y = row[10:].strip()  # operating rules
+                    data_list.append([x, y])
+
+                varrule_data = pd.DataFrame(
+                    data_list, columns=["Time", "Operating Rules"]
+                )
+                varrule_data = varrule_data.set_index("Time")
+                varrule_data = varrule_data["Operating Rules"]
+                self.time_varrule_data = varrule_data
+
+    def _get_gates(self, ngates, block, gate_row):
+        gates = []
+
+        if self.control_method == "TIME":
+            for gate in range(ngates):
+                nrows = int(split_10_char(block[gate_row + 1])[0])
+                data_list = []
+                for row in block[gate_row + 2 : gate_row + 2 + nrows]:
+                    row_split = split_10_char(f"{row:<20}")
+                    x = _to_float(row_split[0])  # time
+                    y = _to_float(row_split[1])  # opening
+                    data_list.append([x, y])
+
+                gate_data = pd.DataFrame(data_list, columns=["Time", "Opening"])
+                gate_data = gate_data.set_index("Time")
+                gate_data = gate_data["Opening"]
+
+                gates.append(gate_data)
+
+                gate_row += 2 + nrows
+
+            self._last_gate_row = gate_row
+
+            return gates
+
+        elif self.control_method == "LOGICAL":
+            for gate in range(ngates):
+                nrows = int(split_10_char(block[gate_row + 1])[0])
+                data_list = []
+                for row in block[gate_row + 2 : gate_row + 2 + nrows]:
+                    row_split = split_10_char(f"{row:<30}")
+                    x = _to_float(row_split[0])  # time
+                    y = row_split[1]  # mode
+                    z = _to_float(row_split[2])  # opening
+                    data_list.append([x, y, z])
+
+                gate_data = pd.DataFrame(data_list, columns=["Time", "Mode", "Opening"])
+                gate_data = gate_data.set_index("Time")
+
+                gates.append(gate_data)
+
+                gate_row += 2 + nrows
+
+            self._last_gate_row = gate_row
+
+            return gates
+
+    def _get_logical_rules(self, nrules, block, rule_row):
+        rules = []
+        rules_recorded = 0
+        rule_logic = []
+        rule_dict = {}
+        nl = "\n"
+        while rules_recorded < nrules:
+            if block[rule_row].strip().upper().endswith(("END", "ENDIF")):
+                rule_logic.append(block[rule_row])
+                rule_dict["logic"] = f"{nl.join(rule_logic)}"
+                rules.append(rule_dict)
+                rule_logic = []
+                rule_dict = {}
+                rules_recorded += 1
+            elif len(rule_dict) == 0:
+                rule_dict = {"name": block[rule_row].strip()}
+            else:
+                rule_logic.append(block[rule_row])
+            rule_row += 1
+
+        self._last_rule_row = rule_row
+
+        return rules
