+++ floodmodeller_api/units/structures.py
@@ -653,63 +653,7 @@ class SLUICE(Unit):
             self.gates = self._get_gates(self.ngates, block, gate_row=7)
 
         elif self.control_method == "LOGICAL":
-            logical_params = split_10_char(block[6])
-            self.max_movement_rate = logical_params[1]
-            self.max_setting = logical_params[2]
-            self.min_setting = logical_params[3]
-
-            self.gates = self._get_gates(self.ngates, block, gate_row=7)
-
-            rule_params = split_10_char(block[self._last_gate_row + 1])
-            self.nrules = int(rule_params[0])
-            self.rule_sample_time = _to_float(rule_params[1])
-
-            self.rules = self._get_logical_rules(
-                self.nrules, block, self._last_gate_row + 2
-            )
-
-            # Get time rule data set
-
-            nrows = int(split_10_char(block[self._last_rule_row + 1])[0])
-            data_list = []
-            for row in block[self._last_rule_row + 2 : self._last_rule_row + 2 + nrows]:
-                row_split = split_10_char(f"{row:<20}")
-                x = _to_float(row_split[0])  # time
-                y = row[10:].strip()  # operating rules
-                data_list.append([x, y])
-            self._last_time_row = self._last_rule_row + nrows + 1
-
-            rule_data = pd.DataFrame(data_list, columns=["Time", "Operating Rules"])
-            rule_data = rule_data.set_index("Time")
-            rule_data = rule_data["Operating Rules"]
-            self.time_rule_data = rule_data
-
-            #VARRULES (not always necessary)
-
-            if self._last_time_row + 1 < len(block):
-                if block[self._last_time_row + 1].strip() == "VARRULES":
-                    varrule_params = split_10_char(block[self._last_time_row + 2])
-                    self.nvarrules = int(varrule_params[0])
-                    self.varrule_sample_time = _to_float(rule_params[1])
-
-                    self.varrules = self._get_logical_rules(
-                        self.nvarrules, block, self._last_time_row + 3
-                    )
-
-                    # Get time rule data set
-
-                    var_nrows = int(split_10_char(block[self._last_rule_row + 1])[0])
-                    data_list = []
-                    for row in block[self._last_rule_row + 2 : self._last_rule_row + 2 + var_nrows]:
-                        row_split = split_10_char(f"{row:<20}")
-                        x = _to_float(row_split[0])  # time
-                        y = row[10:].strip()  # operating rules
-                        data_list.append([x, y])
-                    
-                    varrule_data = pd.DataFrame(data_list, columns=["Time", "Operating Rules"])
-                    varrule_data = varrule_data.set_index("Time")
-                    varrule_data = varrule_data["Operating Rules"]
-                    self.time_varrule_data = varrule_data    
+            self._read_logic(block)
 
         else:
             self._raw_extra_lines = block[6:]
@@ -827,77 +771,6 @@ class SLUICE(Unit):
 
         return block
 
-    def _get_gates(self, ngates, block, gate_row):
-        gates = []
-
-        if self.control_method == "TIME":
-            for gate in range(ngates):
-                nrows = int(split_10_char(block[gate_row + 1])[0])
-                data_list = []
-                for row in block[gate_row + 2 : gate_row + 2 + nrows]:
-                    row_split = split_10_char(f"{row:<20}")
-                    x = _to_float(row_split[0])  # time
-                    y = _to_float(row_split[1])  # opening
-                    data_list.append([x, y])
-
-                gate_data = pd.DataFrame(data_list, columns=["Time", "Opening"])
-                gate_data = gate_data.set_index("Time")
-                gate_data = gate_data["Opening"]
-
-                gates.append(gate_data)
-
-                gate_row += 2 + nrows
-
-            self._last_gate_row = gate_row
-
-            return gates
-
-        elif self.control_method == "LOGICAL":
-            for gate in range(ngates):
-                nrows = int(split_10_char(block[gate_row + 1])[0])
-                data_list = []
-                for row in block[gate_row + 2 : gate_row + 2 + nrows]:
-                    row_split = split_10_char(f"{row:<30}")
-                    x = _to_float(row_split[0])  # time
-                    y = row_split[1]  # mode
-                    z = _to_float(row_split[2])  # opening
-                    data_list.append([x, y, z])
-
-                gate_data = pd.DataFrame(data_list, columns=["Time", "Mode", "Opening"])
-                gate_data = gate_data.set_index("Time")
-
-                gates.append(gate_data)
-
-                gate_row += 2 + nrows
-
-            self._last_gate_row = gate_row
-
-            return gates
-
-    def _get_logical_rules(self, nrules, block, rule_row):
-        rules = []
-        rules_recorded = 0
-        rule_logic = []
-        rule_dict = {}
-        nl = "\n"
-        while rules_recorded < nrules:
-            if block[rule_row].strip().upper().endswith(("END", "ENDIF")):
-                rule_logic.append(block[rule_row])
-                rule_dict["logic"] = f"{nl.join(rule_logic)}"
-                rules.append(rule_dict)
-                rule_logic = []
-                rule_dict = {}
-                rules_recorded += 1
-            elif len(rule_dict) == 0:
-                rule_dict = {"name": block[rule_row].strip()}
-            else:
-                rule_logic.append(block[rule_row])
-            rule_row += 1
-
-        self._last_rule_row = rule_row
-
-        return rules
-
 
 class ORIFICE(Unit):
     """Class to hold and process ORIFICE unit type
@@ -992,7 +865,6 @@ class ORIFICE(Unit):
         surcharged_flow=1.0,
         modular_limit=0.7,
     ):
-
         for param, val in {
             "name": name,
             "flapped": flapped,
@@ -1087,7 +959,6 @@ class SPILL(Unit):
         modular_limit=0.9,
         data=None,
     ):
-
         for param, val in {
             "name": name,
             "ds_label": ds_label,
@@ -1192,7 +1063,6 @@ class RNWEIR(Unit):
         weir_breadth=0.0,
         weir_elevation=0.0,
     ):
-
         for param, val in {
             "name": name,
             "comment": comment,
@@ -1253,7 +1123,6 @@ class CRUMP(Unit):
         self.downstream_crest_height = _to_float(params2[1])
 
     def _write(self):
-
         """Function to write a valid CRUMP block"""
         _validate_unit(self)
         header = "CRUMP " + self.comment
@@ -1296,7 +1165,6 @@ class CRUMP(Unit):
         us_remote_label="",
         ds_remote_label="",
     ):
-
         for param, val in {
             "name": name,
             "comment": comment,
@@ -1420,7 +1288,6 @@ class FLAT_V_WEIR(Unit):
         coriolis_coefficient=1.2,
         bank_top_elevation=0.0,
     ):
-
         for param, val in {
             "name": name,
             "comment": comment,
@@ -1576,7 +1443,6 @@ class RESERVOIR(Unit):  # NOT CURRENTLY IN USE
         lat3="",
         lat4="",
     ):
-
         for param, val in {
             "name": name,
             "comment": comment,
@@ -1691,7 +1557,6 @@ class OUTFALL(Unit):
         surcharged_flow=1.0,
         modular_limit=0.7,
     ):
-
         for param, val in {
             "name": name,
             "flapped": flapped,
